[#39] move: BaseEntity common 패키지로 이동
