+++ src/main/java/bc1/gream/domain/order/entity/Sell.java
@@ -1,6 +1,6 @@
 package bc1.gream.domain.order.entity;
 
-import bc1.gream.domain.model.BaseEntity;
+import bc1.gream.domain.common.model.BaseEntity;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.user.entity.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
