+++ src/main/java/bc1/gream/domain/user/entity/Coupon.java
@@ -1,6 +1,6 @@
 package bc1.gream.domain.user.entity;
 
-import bc1.gream.domain.model.BaseEntity;
+import bc1.gream.domain.common.model.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
